Make react-router work
