+++ package.json
@@ -10,6 +10,7 @@
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "^3.3.0",
     "sass": "^1.25.0"
   },
