+++ src/App.jsx
@@ -1,52 +1,24 @@
-import React from 'react';
-import BigButton from './BigButton';
-import GitInfo from 'react-git-info/macro';
-import packageJson from '../package.json';
+import React, { Component } from 'react';
+import { HashRouter as Router, Switch, Route } from 'react-router-dom';
+import Home from './Home';
 
-import './scss/common.scss';
-import './scss/App.scss';
-
-const links = [
-  {
-    text: 'GitHub',
-    url: 'https://github.com/AbhyudayaSharma',
-  },
-  {
-    text: 'LinkedIn',
-    url: 'https://www.linkedin.com/in/abhyudaya-sharma/',
-  },
-];
-
-function App() {
-  const gitInfo = GitInfo();
-  const buttons = links.map((link, index) => <BigButton key={index} {...link}/>);
-
-  return (
-    <div className="App">
-      <div className="App-header">
-        <h1 className='unselectable'>
-          {packageJson.author.name}
-        </h1>
-      </div>
-      <div className="App-body">
-        {buttons}
-      </div>
-      <div className='App-footer unselectable'>
-        <p>
-          This page was built and deployed from the commit&nbsp;
-          <a href={`${packageJson.repository.url}/commit/${gitInfo.commit.hash}`}
-            className='App-link selectable'>
-            <code>{gitInfo.commit.shortHash}</code>
-          </a>
-          <br/>
-          Fork this repository on&nbsp;
-          <a href={packageJson.repository.url} className='selectable App-link'>
-            GitHub
-          </a>
-        </p>
-      </div>
-    </div>
-  );
+class App extends Component {
+  render() {
+    return (
+      <Router>
+        <Switch>
+          <Route exact path="/">
+            <Home/>
+          </Route>
+          <Route exact path="/blog">
+            <h1>
+              Hello world!
+            </h1>
+          </Route>
+        </Switch>
+      </Router>
+    );
+  }
 }
 
 export default App;
